+++ changelog.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 3.3.0
+
+#### Changes
+- Update Frollo Host API version to 2.5
+- Goals enhancements
+- Send application version in headers
+
 ### 3.2.0
 
 #### Changes